feat(desktop): remove notification step from onboarding, auto-request on first send
Strip the desktop notifications card from the onboarding SetupStep — notifications
now default to enabled for new users and the OS permission prompt fires organically
on first notification delivery via the feed notification hook.

- Flip desktopEnabled default to true in DEFAULT_NOTIFICATION_SETTINGS
- Add auto-permission-request in useFeedDesktopNotifications (with denial → setDesktopEnabled(false))
- Add module-level promise dedup on requestDesktopNotificationAccess to prevent concurrent OS prompts
- Extract useFeedDesktopNotifications to its own file to stay under file size limit
- Remove notification card, types, props, and callbacks from onboarding flow
- Update E2E tests to expect new default state
- Update Settings copy to reflect new default
