+++ desktop/src/features/notifications/use-feed-desktop-notifications.ts
@@ -1,0 +1,198 @@
+import * as React from "react";
+
+import {
+  resolveUserLabel,
+  truncatePubkey,
+  type UserProfileLookup,
+} from "@/features/profile/lib/identity";
+import type { FeedItem, HomeFeedResponse } from "@/shared/api/types";
+import {
+  collectHomeAlertItems,
+  eligibleFeedNotificationItems,
+  notificationBody,
+  notificationTitle,
+} from "./lib/feed";
+import {
+  getDesktopNotificationPermissionState,
+  requestDesktopNotificationAccess,
+  sendDesktopNotification,
+} from "./lib/desktop";
+import { playNotificationSound } from "./lib/sound";
+import type { NotificationSettings } from "./hooks";
+
+const HOME_FEED_SEEN_STORAGE_KEY = "sprout-home-feed-seen.v1";
+const HOME_FEED_SEEN_MAX_ITEMS = 500;
+
+function homeFeedSeenStorageKey(pubkey: string) {
+  return `${HOME_FEED_SEEN_STORAGE_KEY}:${pubkey}`;
+}
+
+export function readStoredSeenFeedIds(pubkey: string): string[] {
+  if (typeof window === "undefined" || pubkey.length === 0) {
+    return [];
+  }
+
+  const rawValue = window.localStorage.getItem(homeFeedSeenStorageKey(pubkey));
+  if (!rawValue) {
+    return [];
+  }
+
+  try {
+    const parsed = JSON.parse(rawValue);
+    if (!Array.isArray(parsed)) {
+      return [];
+    }
+
+    return parsed
+      .filter((value): value is string => typeof value === "string")
+      .slice(-HOME_FEED_SEEN_MAX_ITEMS);
+  } catch {
+    return [];
+  }
+}
+
+export function writeStoredSeenFeedIds(pubkey: string, ids: string[]) {
+  if (typeof window === "undefined" || pubkey.length === 0) {
+    return;
+  }
+
+  window.localStorage.setItem(
+    homeFeedSeenStorageKey(pubkey),
+    JSON.stringify(ids.slice(-HOME_FEED_SEEN_MAX_ITEMS)),
+  );
+}
+
+export function useFeedDesktopNotifications(
+  feed: HomeFeedResponse | undefined,
+  pubkey: string | undefined,
+  settings: NotificationSettings,
+  setDesktopEnabled: (enabled: boolean) => Promise<boolean>,
+  profiles?: UserProfileLookup,
+) {
+  const normalizedPubkey = pubkey?.trim().toLowerCase() ?? "";
+  const seenItemIdsRef = React.useRef<Set<string>>(
+    new Set(readStoredSeenFeedIds(normalizedPubkey)),
+  );
+  const hasAutoRequestedRef = React.useRef(false);
+
+  React.useEffect(() => {
+    seenItemIdsRef.current = new Set(readStoredSeenFeedIds(normalizedPubkey));
+    hasAutoRequestedRef.current = false;
+  }, [normalizedPubkey]);
+
+  const autoRequestPermissionIfNeeded = React.useEffectEvent(async () => {
+    if (hasAutoRequestedRef.current) {
+      return;
+    }
+
+    const currentPermission = await getDesktopNotificationPermissionState();
+    if (currentPermission !== "default") {
+      return;
+    }
+
+    hasAutoRequestedRef.current = true;
+    const result = await requestDesktopNotificationAccess();
+    if (result !== "granted") {
+      void setDesktopEnabled(false);
+    }
+  });
+
+  const deliverFeedNotification = React.useEffectEvent(
+    async (item: FeedItem, senderName?: string) => {
+      const didSend = await sendDesktopNotification({
+        body: notificationBody(item),
+        target: {
+          channelId: item.channelId,
+          channelName: item.channelName,
+          content: item.content,
+          createdAt: item.createdAt,
+          eventId: item.id,
+          kind: item.kind,
+          pubkey: item.pubkey,
+        },
+        title: notificationTitle(item, senderName),
+      });
+
+      if (didSend && settings.soundEnabled) {
+        playNotificationSound();
+      }
+    },
+  );
+
+  React.useEffect(() => {
+    if (!feed) {
+      return;
+    }
+
+    // Wait for sender profiles to load so notification titles include names.
+    // The first-load seed below marks all current items as seen, so we must
+    // defer it until profiles are available — otherwise items get marked seen
+    // before we can dispatch notifications with sender names.
+    if (profiles === undefined) {
+      return;
+    }
+
+    const currentFeedItems = collectHomeAlertItems(feed);
+
+    // Guard: empty seen set + populated feed means first load or cleared
+    // storage. Seed the seen set without notifying to prevent a flood.
+    if (seenItemIdsRef.current.size === 0 && currentFeedItems.length > 0) {
+      seenItemIdsRef.current = new Set(currentFeedItems.map((item) => item.id));
+      writeStoredSeenFeedIds(normalizedPubkey, [...seenItemIdsRef.current]);
+      return;
+    }
+
+    const nextSeenItemIds = new Set(seenItemIdsRef.current);
+    const newItems = settings.desktopEnabled
+      ? eligibleFeedNotificationItems(feed, {
+          mentions: settings.mentions,
+          needsAction: settings.needsAction,
+        }).filter((item) => !nextSeenItemIds.has(item.id))
+      : [];
+
+    for (const item of currentFeedItems) {
+      nextSeenItemIds.add(item.id);
+    }
+
+    // Prevent unbounded growth — keep only the most recent entries.
+    if (nextSeenItemIds.size > HOME_FEED_SEEN_MAX_ITEMS) {
+      const excess = nextSeenItemIds.size - HOME_FEED_SEEN_MAX_ITEMS;
+      let removed = 0;
+      for (const id of nextSeenItemIds) {
+        if (removed >= excess) break;
+        nextSeenItemIds.delete(id);
+        removed++;
+      }
+    }
+
+    seenItemIdsRef.current = nextSeenItemIds;
+    writeStoredSeenFeedIds(normalizedPubkey, [...nextSeenItemIds]);
+
+    if (newItems.length > 0) {
+      void autoRequestPermissionIfNeeded();
+    }
+
+    for (const item of newItems) {
+      const resolvedLabel = profiles
+        ? resolveUserLabel({
+            pubkey: item.pubkey,
+            profiles,
+            preferResolvedSelfLabel: true,
+          })
+        : undefined;
+      // Only use real display names, not truncated pubkey fallbacks.
+      const senderName =
+        resolvedLabel && resolvedLabel !== truncatePubkey(item.pubkey)
+          ? resolvedLabel
+          : undefined;
+      void deliverFeedNotification(item, senderName);
+    }
+  }, [
+    feed,
+    normalizedPubkey,
+    profiles,
+    settings.desktopEnabled,
+    settings.mentions,
+    settings.needsAction,
+  ]);
+}
